// app/src/prediction-market/enums/market-participation.enums.ts
export enum PredictionMarketTradeModesEnum {
  BUY = 'buy',
  SELL = 'sell',
}

export enum PredictionMarketParticipationModesEnum {
  BUY = 'buy',
  SELL = 'sell',
  PROVIDE_LIQUIDITY = 'provide-liquidity',
}

export enum PredictionMarketParticipationResultsEnum {
  WON = 'won',
  LOST = 'lost',
  NOT_RESOLVED = 'not-resolved',
  UNKNOWN = 'unknown',
  NONE = 'none',
}

export enum PredictionMarketParticipationSortByOptionsEnum {
  DATE = 'date',
  AMOUNT = 'amount',
  PAYMENT = 'payment',
  MARKET = 'market',
}